+++ SwiftStomp/Classes/Enumurations/StompAckMode.swift
@@ -1,0 +1,13 @@
+//
+//  StompAckMode.swift
+//  Pods
+//
+//  Created by Ahmad Daneshvar on 5/16/24.
+//
+
+public enum StompAckMode : String{
+    case clientIndividual = "client-individual"
+    case client = "client"
+    case auto = "auto"
+}
+
