+++ SwiftStomp/Classes/Enumurations/StompConnectType.swift
@@ -1,0 +1,12 @@
+//
+//  StompConnectType.swift
+//  Pods
+//
+//  Created by Ahmad Daneshvar on 5/16/24.
+//
+
+public enum StompConnectType{
+    case toSocketEndpoint
+    case toStomp
+}
+
