+++ SwiftStomp/Classes/Enumurations/StompConnectionStatus.swift
@@ -1,0 +1,13 @@
+//
+//  StompConnectionStatus.swift
+//  Pods
+//
+//  Created by Ahmad Daneshvar on 5/16/24.
+//
+
+public enum StompConnectionStatus{
+    case connecting
+    case socketDisconnected
+    case socketConnected
+    case fullyConnected
+}
