+++ SwiftStomp/Classes/Enumurations/StompDisconnectType.swift
@@ -1,0 +1,12 @@
+//
+//  StompDisconnectType.swift
+//  Pods
+//
+//  Created by Ahmad Daneshvar on 5/16/24.
+//
+
+public enum StompDisconnectType{
+    case fromSocket
+    case fromStomp
+}
+
