+++ SwiftStomp/Classes/Enumurations/StompErrorType.swift
@@ -1,0 +1,11 @@
+//
+//  StompErrorType.swift
+//  Pods
+//
+//  Created by Ahmad Daneshvar on 5/16/24.
+//
+
+public enum StompErrorType{
+    case fromSocket
+    case fromStomp
+}
