+++ SwiftStomp/Classes/Enumurations/StompLogType.swift
@@ -1,0 +1,12 @@
+//
+//  StompLogType.swift
+//  Pods
+//
+//  Created by Ahmad Daneshvar on 5/16/24.
+//
+
+internal enum StompLogType : String{
+    case info = "INFO"
+    case socketError = "SOCKET ERROR"
+    case stompError = "STOMP ERROR"
+}
